// wls-gui-wahllokalsystem/src/api/wls-clients/generated-vorfaelleundvorkommnisse-api/api.ts
export interface EreignisDTO {
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'beschreibung'?: string;
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'uhrzeit'?: string;
    /**
     * 
     * @type {string}
     * @memberof EreignisDTO
     */
    'ereignisart': EreignisDTOEreignisartEnum;
}

export const EreignisDTOEreignisartEnum = {
    Vorfall: 'VORFALL',
    Vorkommnis: 'VORKOMMNIS'
} as const;

export type EreignisDTOEreignisartEnum = typeof EreignisDTOEreignisartEnum[keyof typeof EreignisDTOEreignisartEnum];

/**
 * 
 * @export
 * @interface EreignisseWriteDTO
 */
export interface EreignisseWriteDTO {
    /**
     * 
     * @type {boolean}
     * @memberof EreignisseWriteDTO
     */
    'keineVorfaelle'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof EreignisseWriteDTO
     */
    'keineVorkommnisse'?: boolean;
    /**
     * 
     * @type {Array<EreignisDTO>}
     * @memberof EreignisseWriteDTO
     */
    'ereigniseintraege'?: Array<EreignisDTO>;
}
/**
 * 
 * @export
 * @interface WahlbezirkEreignisseDTO
 */
export interface WahlbezirkEreignisseDTO {
    /**
     * 
     * @type {string}
     * @memberof WahlbezirkEreignisseDTO
     */
    'wahlbezirkID': string;
    /**
     * 
     * @type {boolean}
     * @memberof WahlbezirkEreignisseDTO
     */
    'keineVorfaelle'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof WahlbezirkEreignisseDTO
     */
    'keineVorkommnisse'?: boolean;
    /**
     * 
     * @type {Array<EreignisDTO>}
     * @memberof WahlbezirkEreignisseDTO
     */
    'ereigniseintraege'?: Array<EreignisDTO>;
}